Revamped Database data heirarchy
now have two objects:
a user, which stores id, and various other information
and a service provider, which stores identifying information as well as
the associated keys for that service.

service specific tokens for any given user are stored in a relation now,
which uses the ids of the relvant user and service for identification.